move models to their own folder
